Load raw PBM scan data as PBMExperiment

This function reads PBM scan data from GenePix Results (GPR) files and returns a single PBMExperiment object containing foreground and optionally background probe intensity data in assays. Each GPR is stored as a column in the object, with metadata specified in the input scans table stored in the colData of the object.

If probe design information is also specified as a PBMDesign object to the optional probes= argument, this information is also stored in the rowData of the returned PBMExperiment.

gpr2PBMExperiment(scans, useMedian = TRUE, filterFlags = TRUE,
  readBackground = TRUE, probes = NULL, ...)

Arguments

scans
a data.frame of GPR scan information with at least a single `gpr' column, and preferrably a second `type' column. See details for more information on the content of these columns.

useMedian
a logical value whether to use median fluorescent intensity for each probe rather than mean fluorescent intensity. (default = TRUE)

filterFlags
a logical value whether to replace intensity values at probes flagged manually or automatically as being low quality (`Bad': -100, `Absent': -75, `Not Found': -50) with NA. (default = TRUE)

readBackground
a logical value whether to also read in probe background intensities. (default = TRUE)

probes
a PBMDesign object containing the probe design used for the data. (default = NULL)

...
optionally, probe design information can be directly specified using design=, probeFilter=, probeTrim= parameters. These will be ignored if probes= is specified.

Value

PBMExperiment object with assays containing intensities from the GPR files specified in the table of scans. The object includes the following assays: * fore - foreground probe intensities, * back - background probe intensities (unless readBackground = FALSE).

Scan metadata included in the data.frame are stored as colData. Probe design information, if specified, is included as rowData.

Details

The primary argument, scans, must be a data.frame of GPR scan metadata. Each row of the data.frame should correspond to a single GPR scan. At a minimum the data.frame must include a single "gpr" column specifying the paths to the corresponding GPR files to be read in.

Additionally, GPR files are parsed differently depending on the file type which must be one of ``Alexa", ``Cy3", or ``Masliner". Both ``Alexa" and ``Cy3" scans are treated as untouched GPR files, while ``Masliner" files are treated as Masliner-processed GPR files. If possible, the scan type of each file should be specified in a "type" column of the scans data.frame. If a "type" column is not included, all GPR files are assumed to be Alexa488 scans unless the file name in the "gpr" column contains ``Masliner".

See also