Read GPR File

Helper function for reading in a single GPR file.

readGPR(gpr_path, gpr_type, useMedian = TRUE, filterFlags = TRUE,
  readBackground = TRUE)

Arguments

gpr_path
path to GPR file.

gpr_type
scan type; one of "Alexa", "Cy3", "Masliner".

useMedian
logical whether to use median fluorescent intensity for each probe rather than mean fluorescent intensity. (default = TRUE)

filterFlags
logical whether to replace intensity values at probes flagged manually or automatically as being low quality. ('Bad': -100, 'Absent': -75, 'Not Found': -50) with NA. (default = TRUE)

readBackground
logical whether to also read in probe background intensities. (default = TRUE)

Value

tibble (data.frame-like) object of a single GPR file with three or four columns: 'Column', 'Row', 'foreground', (optionally) 'background'. ('ID' and 'Name' columns are ignored as these may be incorrect in the GPR file.)

Details

Since the names of foreground and background columns can vary across samples, they are identified using regular expressions. Therefore, if the column names deviate from the expected format ("F.* Mean", "F.* Median". "B.* Mean", "B.* Median"), the function may fail to read the foreground and background intensities.