PBM Preprocess

This is the main wrapper function for the core PBM data pre-processing steps implemented in the upbm package.

upbmPreprocess(
  pe,
  cy3pe,
  cy3refpe,
  assay = SummarizedExperiment::assayNames(pe)[1],
  params = list(),
  skip = vector("character", 0),
  verbose = 1L
)

Arguments

pe
a PBMExperiment object containing Alexa488 intensity data.

cy3pe
an optional PBMExperiment object containing Cy3 intensity data. If corresponding Cy3 data is not available for the Alexa488 scans, this must be set to NULL.

cy3refpe
an optional PBMExperiment object containing Cy3 reference data to use for Cy3 normalization with cy3FitEmpirical (recommended). To perform model-based normalization with cy3FitModel, set to NULL. This option is ignored if cy3pe = NULL.

assay
a numeric index or string specifying the foreground assay in both the Alexa488 and (optional) Cy3 PBMExperiment objects. Note that this assay is only used for the initial background subtraction and Cy3 normalization steps. All subsequent steps will use the default processed assays from the previous steps. (default = SummarizedExperiment::assayNames(pe)[1])

params
a named list specifying any non-default parameters that should be used for preprocessing the data. Only entries in the list exactly matching the name of a preprocessing step are considered valid ("backgroundSubtract", "cy3Normalize", "spatiallyAdjust", "normalizeWithinReplicates", or "normalizeAcrossReplicates"). For each step, non-default parameters must be specified as a list of named parameter values. (default = list()). See below for examples.

skip
a string vector specifying any steps to skip during preprocessing. Strings must exactly match step names as with the params parameter. (default = vector("character", 0))

verbose
an integer value specifying the level of verbosity. Must be one of 0, 1, or 2. If 0, no progress messages are printed. If 1, progress messages are only provided to indicate when each preprocessing step is started and finished. If 2, in addition to the basic progress messages, verbose = TRUE is also specified for all preprocessing functions called internally. (default = 1L)

Value

PBMExperiment containing filtered and normalized Alexa488 probe intensities.

See also

Examples

data(hoxc9alexa, package = "upbmData")
data(hoxc9cy3, package = "upbmData")
data(refcy3_8x60k_v1, package = "upbmAux")

if (FALSE) {
upbmPreprocess(hoxc9alexa, hoxc9cy3, refcy3_8x60k_v1)

upbmPreprocess(hoxc9alexa, NULL, NULL, params = list(backgroundSubtract = FALSE))

upbmPreprocess(hoxc9alexa, hoxc9cy3, refcy3_8x60k_v1,
               params = list(backgroundSubtract = list(assayb = "back")),
               skip = "cy3Normalize")
}